Fix cyclic loading of the modules (#330)
