+++ src/test/scala/me/rexim/morganey/module/ModuleIndexSpec.scala
@@ -1,5 +1,7 @@
 package me.rexim.morganey.module
 
+import me.rexim.morganey.mock._
+
 import org.mockito.Mockito._
 import org.scalatest._
 import org.scalatest.mockito.MockitoSugar
@@ -9,7 +11,7 @@ import java.util.Vector
 import java.net.URL
 import java.nio.charset.StandardCharsets
 
-class ModuleIndexSpec extends FlatSpec with MockitoSugar with Matchers {
+class ModuleIndexSpec extends FlatSpec with Matchers with ClassLoaderMocking {
   def mockMorganeyIndexUrl(modulesInIndex: List[String]): URL = {
     val morganeyIndexContent = modulesInIndex.mkString("\n")
 
@@ -25,8 +27,11 @@ class ModuleIndexSpec extends FlatSpec with MockitoSugar with Matchers {
   it should "return all modules from the index" in {
     import scala.collection.JavaConverters._
 
-    val moduleContainers = List(List("hello.mgn", "world.mgn"), List("a/foo.mgn", "b/bar.mgn"))
-    val expectedModules = moduleContainers.flatMap(identity).map(path => new Module(ResourcePath(path)))
+    val moduleContainers = List(
+      List("hello.mgn", "world.mgn"),
+      List("a/foo.mgn", "b/bar.mgn")
+    )
+    val expectedModules = moduleContainers.flatten.map(path => new Module(ResourcePath(path)))
     val mockedMorganeyIndexUrls = moduleContainers.map(mockMorganeyIndexUrl)
 
     val classLoader = mock[ClassLoader]
